deprecate search_options in favor of search-options; rename response fields and speed up the new call
