+++ chpl/chpl-api/src/main/java/gov/healthit/chpl/web/controller/SearchViewController.java
@@ -39,6 +39,7 @@ import gov.healthit.chpl.domain.CertificationCriterion;
 import gov.healthit.chpl.domain.CriteriaSpecificDescriptiveModel;
 import gov.healthit.chpl.domain.DecertifiedDeveloperResult;
 import gov.healthit.chpl.domain.DescriptiveModel;
+import gov.healthit.chpl.domain.DimensionalData;
 import gov.healthit.chpl.domain.FuzzyChoices;
 import gov.healthit.chpl.domain.KeyValueModel;
 import gov.healthit.chpl.domain.KeyValueModelStatuses;
@@ -939,7 +940,7 @@ public class SearchViewController {
     produces = "application/json; charset=utf-8")
     @CacheControl(policy = CachePolicy.PUBLIC, maxAge = CacheMaxAge.TWELVE_HOURS)
     public @ResponseBody Set<KeyValueModelStatuses> getProductNames() {
-        return dimensionalDataManager.getProductNames();
+        return dimensionalDataManager.getProducts();
     }
 
     @ApiOperation(value = "Get all possible developer names in the CHPL",
@@ -948,7 +949,7 @@ public class SearchViewController {
     produces = "application/json; charset=utf-8")
     @CacheControl(policy = CachePolicy.PUBLIC, maxAge = CacheMaxAge.TWELVE_HOURS)
     public @ResponseBody Set<KeyValueModelStatuses> getDeveloperNames() {
-        return dimensionalDataManager.getDeveloperNames();
+        return dimensionalDataManager.getDevelopers();
     }
 
     @ApiOperation(value = "Get all possible ACBs in the CHPL",
@@ -1216,16 +1217,35 @@ public class SearchViewController {
      * @return a map of all filterable values
      * @throws EntityRetrievalException if an item cannot be retrieved from the db
      */
-    @ApiOperation(value = "Get all search options in the CHPL",
+    @Deprecated
+    @ApiOperation(value = "DEPRECATED. Use /data/search-options instead. Get all search options in the CHPL",
             notes = "This returns all of the other /data/{something} results in one single response.")
     @RequestMapping(value = "/data/search_options", method = RequestMethod.GET,
     produces = "application/json; charset=utf-8")
     @CacheControl(policy = CachePolicy.PUBLIC, maxAge = CacheMaxAge.TWELVE_HOURS)
-    public @ResponseBody SearchableDimensionalData getSearchOptions(
+    public @ResponseBody SearchableDimensionalData getSearchOptionsDeprecated(
             @RequestParam(value = "simple", required = false, defaultValue = "false") final Boolean simple)
                     throws EntityRetrievalException {
 
         return dimensionalDataManager.getSearchableDimensionalData(simple);
+    }
+
+    /**
+     * Returns all of the fields that have a finite set of values and may be used
+     * as filers when searching for listings.
+     * @param simple whether to include data relevant to 2011 listings (2011 edition and NQF numbers)
+     * @return a map of all filterable values
+     * @throws EntityRetrievalException if an item cannot be retrieved from the db
+     */
+    @ApiOperation(value = "Get all search options in the CHPL",
+            notes = "This returns all of the other /data/{something} results in one single response.")
+    @RequestMapping(value = "/data/search-options", method = RequestMethod.GET,
+    produces = "application/json; charset=utf-8")
+    @CacheControl(policy = CachePolicy.PUBLIC, maxAge = CacheMaxAge.TWELVE_HOURS)
+    public @ResponseBody DimensionalData getSearchOptions(
+            @RequestParam(value = "simple", required = false, defaultValue = "false") final Boolean simple)
+                    throws EntityRetrievalException {
+        return dimensionalDataManager.getDimensionalData(simple);
     }
 
     @Deprecated
