+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/caching/PrefetchedCacheLoader.java
@@ -63,7 +63,7 @@ public class PrefetchedCacheLoader {
     @Cacheable(CacheNames.PREFETCHED_PRODUCT_NAMES)
     public Set<KeyValueModelStatuses> loadPrefetchedProductNames() {
         LOGGER.info("Loading prefetched product names");
-        Set<KeyValueModelStatuses> results = precacheableDataManager.getProductNames();
+        Set<KeyValueModelStatuses> results = precacheableDataManager.getProducts();
         LOGGER.info("Completed loading product names.");
         return results;
     }
@@ -73,7 +73,7 @@ public class PrefetchedCacheLoader {
     @Cacheable(CacheNames.PREFETCHED_DEVELOPER_NAMES)
     public Set<KeyValueModelStatuses> loadPrefetchedDeveloperNames() {
         LOGGER.info("Loading prefetched developer names");
-        Set<KeyValueModelStatuses> results = precacheableDataManager.getDeveloperNames();
+        Set<KeyValueModelStatuses> results = precacheableDataManager.getDevelopers();
         LOGGER.info("Completed loading developer names.");
         return results;
     }
